added loading forecast for current user location

+++ app/src/main/java/vkraevskiy/com/simpleweatherapp/core/bridge/NetFacade.java
@@ -3,5 +3,5 @@ package vkraevskiy.com.simpleweatherapp.core.bridge;
 import vkraevskiy.com.simpleweatherapp.core.callback.NetCallback;
 
 public interface NetFacade {
-    void loadForecast(long latitude, long longitude, NetCallback netCallback);
+    void loadForecast(double latitude, double longitude, NetCallback netCallback);
 }
